Fix: Correctly locate the test program to load
A fix for locating the correct `program.so` file has been added. Currently
the course loads the program using `Mollusk::new(&ID, "target/deploy/program.so")`
causing testing to crash. This is because mollusk-svm adds the `.so` file extension in the
within the `load_program_elf` function:

https://github.com/anza-xyz/mollusk/blob/e50cf9fbc69b92f347978bb5528192d90d8a88a3/harness/src/file.rs#L81

the fix for this is removing the `.so` in the `Mollusk::new(&ID, "target/deploy/program)` code.

Modifications to the french version have been made to mirror the English version

Closes #136