fix: Avoid manually moving the dependencies to the dev-dependencies section
Mollusk introduction course manually requires devs to add dependencies
to the `dev-dependencies` section of their `Cargo.toml` file yet they
are already using `cargo add` to add dependencies.

This commit adds a `--dev` flag to each `cargo add <crate-name>` to automatically
add each dependency to the `dev` section of the manifest file.

This add the information to the french course too

Depends on PR #137 since it incorporates a fix from that PR
Closes #138