chore: Hook ``` to generate Codeblock component in MDX
The current approach of forcing code block to be wrapped inside a pair
of Codeblock tags to get copy functionality and a more stylized frame
around code snippets is causing a lot of issue with code editor and AI.

The solution is to hook ``` to directly generate the body of the
Codeblock component, so the output is transparently handled by the MDX
compiler. This approach results in a simpler and more consistent content
production experience, and keeps both editors and AI happy.